requirements.txt
certifi==2019.6.16
chardet==3.0.4
Deprecated==1.2.6
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.0.2
django-dotenv==1.4.2
django-extensions==2.2.1
django-taggit==1.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.10.2
Faker==2.0.1
gunicorn==19.9.0
idna==2.8
numpy==1.17.0
psycopg2==2.8.4
PyGithub==1.43.7
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
whitenoise==4.1.3
wrapt==1.11.2
